定数定義位置の見直し
