+++ src/pages/projEstimate/Tables/QuoteTable.tsx
@@ -6,7 +6,8 @@ import TableHead from '@mui/material/TableHead';
 import TableRow from '@mui/material/TableRow';
 import Paper from '@mui/material/Paper';
 import RenderRows from './RenderRows';
-import { materialsNameList, TFullMaterials } from '../form';
+import { TFullMaterials } from '../form';
+import { materialsNameList } from '../constantDefinition';
 
 export type QuoteTableProps = {
   arrayHelpers: any,
