+++ src/pages/projEstimate/constantDefinition.ts
@@ -11,3 +11,15 @@ export const buzaiListInit = {
   unitPrice: 0,
   price: 0,
 };
+
+export const materialsLabelList = [
+  'number', 'majorItem', 'middleItem', 'element', 'costPrice',
+  'quantity', 'elemProfRate', 'unit', 'tax', 'unitPrice', 'price',
+];
+
+/**
+ * 見積もりアイテムのラベル定義
+ */
+export const materialsNameList = [
+  'No.', '大項目*', '中項目', '部材', '原価*', '数量*', '利益率', '単位', '税(課税 / 非課税)', '単価', '金額',
+];
