+++ src/pages/projEstimate/form.ts
@@ -17,21 +17,10 @@ export type TMaterials = {
   unitPrice: number,
   price: number,
 };
-export const materialsLabelList = [
-  'number', 'majorItem', 'middleItem', 'element', 'costPrice',
-  'quantity', 'elemProfRate', 'unit', 'tax', 'unitPrice', 'price',
-];
 
 export type TFullMaterials = {
   items: Array<TMaterials>,
 };
-
-/**
- * 見積もりアイテムのラベル定義
- */
-export const materialsNameList = [
-  'No.', '大項目*', '中項目', '部材', '原価*', '数量*', '利益率', '単位', '税(課税 / 非課税)', '単価', '金額',
-];
 
 export const initialValues = {
   projId: '',
